+++ include/dlaf/communication/kernels/all_reduce.h
@@ -23,9 +23,9 @@
 #include "dlaf/communication/communicator.h"
 #include "dlaf/communication/message.h"
 #include "dlaf/communication/rdma.h"
-#include "dlaf/communication/with_communication_tile.h"
 #include "dlaf/matrix/tile.h"
 #include "dlaf/sender/transform_mpi.h"
+#include "dlaf/sender/with_temporary_tile.h"
 
 namespace dlaf::comm {
 namespace internal {
