+++ include/dlaf/communication/kernels/reduce.h
@@ -24,11 +24,11 @@
 #include "dlaf/common/pipeline.h"
 #include "dlaf/communication/communicator.h"
 #include "dlaf/communication/message.h"
-#include "dlaf/communication/with_communication_tile.h"
 #include "dlaf/matrix/tile.h"
 #include "dlaf/schedulers.h"
 #include "dlaf/sender/traits.h"
 #include "dlaf/sender/transform_mpi.h"
+#include "dlaf/sender/with_temporary_tile.h"
 
 namespace dlaf::comm {
 namespace internal {
